refactor: update tsconfig configuration
- Reorganize project structure
- Remove unused components
- Rename imports for better readability
