+++ src/entities/todo/model/use-todo.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { TODO_QUERY_KEY } from '@/constants';
-import { TodoService } from '@/services';
+import { TodoService } from '@/entities';
+import { TODO_QUERY_KEY } from '@/shared/constants';
 
 export const useTodo = () => {
     return useQuery({ queryKey: TODO_QUERY_KEY, queryFn: () => TodoService.fetchTodo() });
