+++ src/features/counter/model/counter.ts
