+++ src/features/index.ts
@@ -1,0 +1,1 @@
+export * from '@/features/counter/model/counter';
