+++ src/pages/data-page/index.tsx
@@ -1,6 +1,6 @@
-import { useTodo } from '@/hooks';
+import { useTodo } from '@/entities';
 
-export default function DataPage() {
+export function DataPage() {
     const { data, isLoading } = useTodo();
     if (isLoading) return <p>Загрузка...</p>;
     return <pre className="bg-muted p-2 rounded">{JSON.stringify(data, null, 2)}</pre>;
