+++ src/pages/home-page/index.tsx
@@ -1,3 +1,3 @@
-export default function HomePage() {
+export function HomePage() {
     return <div className="text-xl font-bold">rs-react-vite-template</div>;
 }
