+++ src/pages/index.ts
@@ -1,0 +1,4 @@
+export * from '@/pages/counter-page/index';
+export * from '@/pages/data-page/index';
+export * from '@/pages/form-page/index';
+export * from '@/pages/home-page/index';
