--- src/schemas/index.ts
@@ -1,2 +1,0 @@
-export * from './form';
-export * from './todo';
