+++ src/shared/ui/components/index.ts
@@ -1,0 +1,7 @@
+export * from './nav';
+export * from './ui/button';
+export * from './ui/form';
+export * from './ui/input';
+export * from './ui/label';
+export * from './ui/navigation-menu';
+export * from './ui/navigation-menu';
