+++ src/shared/ui/components/ui/form.tsx
@@ -13,8 +13,8 @@ import {
     useFormState,
 } from 'react-hook-form';
 
-import { Label } from '@/components/ui/label';
-import { cn } from '@/lib/utils';
+import { Label } from '@/shared';
+import { cn } from '@/shared';
 
 const Form = FormProvider;
