+++ src/shared/ui/components/ui/label.tsx
@@ -1,7 +1,7 @@
 import * as LabelPrimitive from '@radix-ui/react-label';
 import * as React from 'react';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/shared';
 
 function Label({ className, ...props }: React.ComponentProps<typeof LabelPrimitive.Root>) {
     return (
