--- src/stores/index.ts
@@ -1,1 +1,0 @@
-export * from './counter';
