+++ tsconfig.json
@@ -4,7 +4,13 @@
     "compilerOptions": {
         "baseUrl": ".",
         "paths": {
-            "@/*": ["./src/*"]
+            "@/*": ["./src/*"],
+            "@/entities/*": ["./src/entities/*"],
+            "@/features/*": ["./src/features/*"],
+            "@/pages/*": ["./src/pages/*"],
+            "@/shared/*": ["./src/shared/*"],
+            "@/widgets/*": ["./src/widgets/*"],
+            "@/app/*": ["./src/app/*"]
         }
     }
 }
